+++ readium/shared/src/test/java/org/readium/r2/shared/publication/LocatorTest.kt
@@ -12,8 +12,11 @@ package org.readium.r2.shared.publication
 import org.json.JSONObject
 import org.junit.Assert.*
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.readium.r2.shared.assertJSONEquals
+import org.robolectric.RobolectricTestRunner
 
+@RunWith(RobolectricTestRunner::class)
 class LocatorTest {
 
     @Test fun `parse {Locator} minimal JSON`() {
@@ -296,6 +299,7 @@ class LocatorTest {
 
 }
 
+@RunWith(RobolectricTestRunner::class)
 class LocatorCollectionTest {
 
     @Test fun `parse {LocatorCollection} minimal JSON`() {