+++ readium/streamer/src/main/java/org/readium/r2/streamer/parser/epub/Metadata.kt
@@ -165,6 +165,7 @@ internal class PubMetadataAdapter(
 
     fun metadata() = Metadata(
         identifier = identifier,
+        conformsTo = setOf(Publication.Profile.EPUB),
         modified = modified,
         published = published,
         languages = languages,