+++ readium/streamer/src/test/java/org/readium/r2/streamer/extensions/FileTest.kt
@@ -11,7 +11,7 @@ package org.readium.r2.streamer.extensions
 
 import org.junit.Test
 import java.io.File
-import kotlin.test.assertEquals
+import org.junit.Assert.assertEquals
 
 class FileTest {