+++ readium/streamer/src/test/java/org/readium/r2/streamer/parser/image/ImageParserTest.kt
@@ -12,17 +12,16 @@ package org.readium.r2.streamer.parser.image
 import kotlinx.coroutines.runBlocking
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
-import org.readium.r2.shared.fetcher.ArchiveFetcher
 import org.readium.r2.shared.fetcher.Fetcher
-import org.readium.r2.shared.fetcher.FileFetcher
+import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.asset.FileAsset
 import org.readium.r2.shared.publication.asset.PublicationAsset
 import org.readium.r2.shared.publication.firstWithRel
 import org.readium.r2.shared.util.archive.DefaultArchiveFactory
 import org.readium.r2.streamer.parseBlocking
 import java.io.File
-import kotlin.test.assertEquals
-import kotlin.test.assertNotNull
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotNull
 
 class ImageParserTest {
 
@@ -37,7 +36,7 @@ class ImageParserTest {
     private fun assetForResource(resource: String): PublicationAsset {
         val path = ImageParserTest::class.java.getResource(resource)?.path
         assertNotNull(path)
-        return FileAsset(File(path))
+        return FileAsset(File(path!!))
     }
 
     private fun fetcherForAsset(asset: PublicationAsset): Fetcher = runBlocking {
@@ -55,10 +54,15 @@ class ImageParserTest {
     }
 
     @Test
+    fun `conformsTo contains the Divina profile`() {
+        assertEquals(setOf(Publication.Profile.DIVINA), parser.parseBlocking(cbzAsset, cbzFetcher)?.manifest?.metadata?.conformsTo)
+    }
+
+    @Test
     fun `readingOrder is sorted alphabetically`() {
         val builder = parser.parseBlocking(cbzAsset, cbzFetcher)
         assertNotNull(builder)
-        val readingOrder = builder.manifest.readingOrder
+        val readingOrder = builder!!.manifest.readingOrder
             .map { it.href.removePrefix("/Cory Doctorow's Futuristic Tales of the Here and Now") }
         assertThat(readingOrder)
             .containsExactly("/a-fc.jpg", "/x-002.jpg", "/x-003.jpg", "/x-004.jpg")
@@ -68,7 +72,7 @@ class ImageParserTest {
     fun `the cover is the first item in the readingOrder`() {
         val builder = parser.parseBlocking(cbzAsset, cbzFetcher)
         assertNotNull(builder)
-        with(builder.manifest.readingOrder) {
+        with(builder!!.manifest.readingOrder) {
             assertEquals(
                 "/Cory Doctorow's Futuristic Tales of the Here and Now/a-fc.jpg",
                 firstWithRel("cover")?.href)
@@ -79,6 +83,6 @@ class ImageParserTest {
     fun `title is based on archive's root directory when any`() {
         val builder = parser.parseBlocking(cbzAsset, cbzFetcher)
         assertNotNull(builder)
-        assertEquals("Cory Doctorow's Futuristic Tales of the Here and Now", builder.manifest.metadata.title)
+        assertEquals("Cory Doctorow's Futuristic Tales of the Here and Now", builder!!.manifest.metadata.title)
     }
 }