+++ test-app/src/main/java/org/readium/r2/testapp/reader/ReaderActivity.kt
@@ -22,9 +22,6 @@ import androidx.lifecycle.ViewModelProvider
 import org.readium.r2.navigator.ExperimentalPresentation
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
-import org.readium.r2.shared.publication.allAreAudio
-import org.readium.r2.shared.publication.allAreBitmap
-import org.readium.r2.shared.util.mediatype.MediaType
 import org.readium.r2.testapp.R
 import org.readium.r2.testapp.databinding.ActivityReaderBinding
 import org.readium.r2.testapp.drm.DrmManagementContract
@@ -61,7 +58,7 @@ open class ReaderActivity : AppCompatActivity() {
 
         if (savedInstanceState == null) {
 
-            if (publication.type == Publication.TYPE.EPUB) {
+            if (publication.conformsTo(Publication.Profile.EPUB)) {
                 val baseUrl = requireNotNull(inputData.baseUrl)
                 readerFragment = EpubReaderFragment.newInstance(baseUrl)
 
@@ -70,9 +67,9 @@ open class ReaderActivity : AppCompatActivity() {
                 }
             } else {
                 val readerClass: Class<out Fragment> = when {
-                    publication.readingOrder.all { it.mediaType == MediaType.PDF } -> PdfReaderFragment::class.java
-                    publication.readingOrder.allAreBitmap -> ImageReaderFragment::class.java
-                    publication.readingOrder.allAreAudio -> AudioReaderFragment::class.java
+                    publication.conformsTo(Publication.Profile.PDF) -> PdfReaderFragment::class.java
+                    publication.conformsTo(Publication.Profile.DIVINA) -> ImageReaderFragment::class.java
+                    publication.conformsTo(Publication.Profile.AUDIOBOOK) -> AudioReaderFragment::class.java
                     else -> throw IllegalArgumentException("Cannot render publication")
                 }
 
@@ -144,6 +141,7 @@ open class ReaderActivity : AppCompatActivity() {
             is ReaderViewModel.Event.Failure -> {
                 Toast.makeText(this, event.error.getUserMessage(this), Toast.LENGTH_LONG).show()
             }
+            else -> {}
         }
     }